added makefile and save_file logic for different files as well as new menu for the beggining, still really buggy

+++ cpps/board.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 
 bool vector_find(const vector<int>& vec, int value){
-    for(int i = 0; i < vec.size(); i++){
+    for(long unsigned int i = 0; i < vec.size(); i++){
         if(vec[i] == value){
             return true;
         }
